moved help contents into its own folder to reduce the amount of files in the data folder
