+++ tests/e2e/data/help-contents/archetype-contents.ts
