+++ tests/e2e/data/help-contents/contract-types-contents.ts
