+++ tests/e2e/data/help-contents/typechecker-simulator-contents.ts
